+++ packages/dash-table/demo/App.js
@@ -25,6 +25,7 @@ class App extends Component {
                     deletable: true,
                 //     type: 'dropdown'
                 })),
+                content_style: 'grow',
                 editable: true,
                 sorting: true,
                 n_fixed_rows: 4,